requirements.txt
Flask
Flask-Cors
azure-storage-blob
gunicorn # For production deployment on App Service
moviepy
werkzeug
psycopg2-binary